Implemented file metrics
Created a KotlinProcessor which task is to extract metrics from
individual kotlin files. The KotlinProcessor is integrated in
KotlinSensor.

Implemented a test for DetektSensor and added some kotlin files in
resources to be used for this and future tests. Also added a couple of
dependencies for testing.

Refactored KotlinSyntax to use KtFile instead of extracting it each
time.